Initial commit of cbexigen
This initial state of the codec generator successfully creates code for the
appHandshake, DIN 70121 and ISO 15118-2 protocols.

Some grammars are incorrect, therefore not all messages can be encoded
correctly or decoded successfully. This is work in progress.

Handling of ISO 15118-20 is not complete, due to its somewhat different
structure, but also work in progress.

Co-authored-by: aw <aw@pionix.de>
Co-authored-by: Christoph Hauschke <christoph.hauschke@chargebyte.com>
Co-authored-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>
Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>