SchemaAnalyzer: fix order of expanded abstract particles
Particles which are referenced by their abstract type, but obviously expand
to all their non-abstract (concrete) implementations, allow for each of
these implementations to appear as an event.

The order in which the complete list of possible particle events was
inserted was incorrect. They are now inserted at the original position in
the list.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>