ExiBaseCoderCode: recursively apply subsequent grammars
This probably only applies to ISO 15118-2 ConsumptionCost in
SalesTariffEntryType, within the scope of DIN 70121 and ISO 15118-2.

Ironically, this has no positive or negative effect for these protocols, as
the position of ConsumptionCost within SalesTariffEntryType was fixed with
the fixes for the order of insertion of abstract and real particles, and is
now at the end of its element anyway. Its next grammar decisions are still
incorrect, as before.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>