fix encoding and decoding of arrays of unsignedShort
Add special handling for arrays of unsignedShort (uint16).

Previously (DIN 70121 and ISO 15118-2), only arrays of complex elements and
enums had special handling in the code creation process. In ISO 15118-20,
there is now also ServiceIDListType, which is an array of ServiceID, which
is an unsignedShort.

Consider this a hack, as arrays of other basic types are not yet supported.
Note that, as hinted above, none of these other arrays appear in the known
V2G schemas.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>