ExiBaseCoderCode: move grammar details for 'ANY' particles to end of list, add dummies
The 'ANY' order change (moving these details to the end) practically only
affects TransformType, where '##other' is listed before 'XPath', but needs
to be after it.

The grammar details for 'END' are now also placed at the end of the list, as
previously done in move_end_element_to_end_of_list().

Additional 'ANY' grammars are added _after_ 'END' in case their originals
appear before, and the ones before are marked as generic "dummies".

If no 'END' exists, no dummies are currently created. The actual
requirement for this is not known. (This would only be the case, if any
'ANY' particle were non-optional, which is not the case in V2G - except for
one case caused by a bug in our analyzer.)

The 'ANY' implementations may need to be fine-tuned for "lax". The actual
requirement for this is not known.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>