Add some minor fixes
The entry in error.log was corrected because the comparison of the increased string lengths was no longer correct.

The definition for EXI_STRING_MAX_LEN is increased to 60. Some strings below Signature in the header are longer than 50 bytes and caused the error EXI_ERROR__CHARACTER_BUFFER_TOO_SMALL (-111).

Signed-off-by: chausGit <christoph.hauschke@chargebyte.com>