config, typeDefinitions: allow up to 4 Reference items
The size of the Reference array was previously reduced to 1, which is not
sufficient for PnC. According to ISO 15118-2 [V2G2-909], 4 items must be
supported.

According to ISO 15118-20, there is no restriction, therefore, this may need
to be revisited, and the limit lifted.

For DIN, the restriction remains 1, as PnC is unused.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>