DatatypeHeader: allow elements of same type in base struct
Convert the intermediate dict, as passed to Jinja, to a list of tuples.

This fixes ISO 15118-20 ACDP missing variable declarations for
ACDP_ConnectReq and ACDP_ConnectRes.

Previously, a python dict with Elements' prefixed_type key was used to list
all base elements to declare. When several elements are of the same type,
these overwrote each other in the dict, whereby their variables weren't
created.

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>