low-level EXI codec: fix left-shift for large integers
The left-hand side of a left-shift operation only gets promoted up to "int",
if too small to support the number of bits created. This leads to
truncation in case of resulting values with more than 32 significant bits,
which are expected in a 64-bit value decoder.

Move the typecast to the left-shifted variable. Do the same for the 32-bit
case, for consistency.

Fixes https://github.com/EVerest/cbexigen/issues/49

Signed-off-by: Moritz Barsnick <moritz.barsnick@chargebyte.com>