Fix the standard fragment elements in the config and add dict for ambiguous elements
The list of standard fragment elements that can be signed now corresponds
to the specifications from the respective standard.

In ISO 15118-2 there is an element with an ambiguous type but the same name.
To create uniqueness, there is now a list in which the type to be used can be
assigned to the element.

Signed-off-by: chausGit <christoph.hauschke@chargebyte.com>