Fix the missing end fragment in the en- and decoder
An end fragment value was missing in the encoder and decoder.
This value is +1 to the total number of possible fragment elements.

A new error code has also been added.

In addition, a note is generated for the DIN and ISO-2 fragment encoder and decoder.
This contains information for checking the signature of the EMAIDType.

Signed-off-by: chausGit <christoph.hauschke@chargebyte.com>